tools/pyproject.toml
[project]
name = "npl-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
npl-loader = "npl_loader:main"

[tool.setuptools]
py-modules = ["npl_loader"]
